Add Safeguard to Destroy Method

+++ app/Http/Controllers/UserController.php
@@ -82,6 +82,9 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        if ($user->id == 1) {
+            return redirect()->route('users.index');
+        }
         $user->delete();
         return redirect()->route('users.index');
     }
